+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-04-11 21:25:42
-// Commit: feat(storage): add bi-directional event bus handler for storage
+// Updated at: 2026-04-12 09:38:31
+// Commit: fix(storage): fix memory leak by cancelling storage subscription
 
 class storageComponentState {
-  final String commitRef = '320';
-  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
-  final DateTime timestamp = DateTime.parse('2026-04-11T21:25:42+05:00');
+  final String commitRef = '325';
+  final String lastMessage = 'fix(storage): fix memory leak by cancelling storage subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-12T09:38:31+05:00');
 }