+++ CHANGELOG.md
@@ -326,3 +326,4 @@
 - [2026-04-11 22:27] fix(analytics): prevent double tap duplicate action in analytics
 - [2026-04-11 23:03] fix(security): prevent double tap duplicate action in security
 - [2026-04-12 09:05] fix(forms): prevent double tap duplicate action in forms
+- [2026-04-12 09:38] fix(storage): fix memory leak by cancelling storage subscription